EasyMCP/MCPDaily/requirements.txt
# MCPDaily Requirements
# This file lists all dependencies for the MCPDaily application

# Core framework (already installed from .whl, included for reference)
# mcplearn-mcp==0.1.0

# Web framework (required by HTTPServer transport)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# HTTP requests (for Weather and News tools)
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# Optional: For better timezone support
tzdata>=2023.3
